Added theme mode "system". Login page respects theme mode.
